refactor after review
