+++ src/main/java/ru/praktikum_services/qa_scooter/courier/CourierAsserts.java
@@ -1,5 +1,6 @@
-package Courier;
+package ru.praktikum_services.qa_scooter.courier;
 
+import io.qameta.allure.Step;
 import io.restassured.response.ValidatableResponse;
 
 import static org.hamcrest.CoreMatchers.equalTo;
@@ -7,6 +8,7 @@ import static org.hamcrest.CoreMatchers.is;
 import static org.hamcrest.Matchers.greaterThan;
 
 public class CourierAsserts {
+    @Step("Валидация статус кода = 201. Проверка тела ответа ok = true")
     public void assertCourierCreatePositiveNewCourier(ValidatableResponse response) {
         response
                 .assertThat()
@@ -14,6 +16,7 @@ public class CourierAsserts {
                 .body("ok", is(true));
     }
 
+    @Step("Валидация статус кода = 409. Проверка тела ответа message = Этот логин уже используется. Попробуйте другой.")
     public void assertCourierCreateNegativeAlreadyExist(ValidatableResponse response) {
         response
                 .assertThat()
@@ -21,6 +24,7 @@ public class CourierAsserts {
                 .body("message", equalTo("Этот логин уже используется. Попробуйте другой."));
     }
 
+    @Step("Валидация статус кода = 400. Проверка тела ответа message = Недостаточно данных для создания учетной записи")
     public void assertCreateCourierNegativeInvalidBody(ValidatableResponse response) {
         response
                 .assertThat()
@@ -28,6 +32,7 @@ public class CourierAsserts {
                 .body("message", equalTo("Недостаточно данных для создания учетной записи"));
     }
 
+    @Step("Валидация статус кода = 200. Проверка тела ответа id больше 0")
     public void assertNewCourierAutentificationCorrect(ValidatableResponse response) {
         response
                 .assertThat()
@@ -35,6 +40,7 @@ public class CourierAsserts {
                 .body("id", greaterThan(0));
     }
 
+    @Step("Валидация статус кода = 400. Проверка тела ответа message = Недостаточно данных для входа")
     public void assertNewCourierAutentificationWithOutMandatoryParams(ValidatableResponse response) {
         response
                 .assertThat()
@@ -42,6 +48,7 @@ public class CourierAsserts {
                 .body("message", equalTo("Недостаточно данных для входа"));
     }
 
+    @Step("Валидация статус кода = 404. Проверка тела ответа message = Учетная запись не найдена")
     public void assertNewCourierAutentificationWithNonExistParam(ValidatableResponse response) {
         response
                 .assertThat()
