+++ src/main/java/ru/praktikum_services/qa_scooter/orders/OrderData.java
@@ -1,4 +1,4 @@
-package Orders;
+package ru.praktikum_services.qa_scooter.orders;
 
 import java.util.List;
